refactor: combine hooks into single file
