+++ src/provide.svelte.ts
@@ -1,4 +1,4 @@
-import type { Store } from "@laserware/stasis";
+import type { Store } from "@reduxjs/toolkit";
 
 import { getStateContextKey, storeContextKey } from "./context.js";
