+++ tsconfig.json
@@ -14,6 +14,7 @@
     "sourceMap": false,
     "strict": true,
     "strictNullChecks": true,
+    "stripInternal": true,
     "target": "ESNext",
     "verbatimModuleSyntax": true
   },
